+++ src/main/java/de/samply/explorer/TaggedLinesExplorer.java
@@ -1,16 +1,11 @@
 package de.samply.explorer;
 
-import de.samply.explorer.ExplorerException;
-import de.samply.explorer.ExplorerImpl;
-import de.samply.explorer.Pivot;
 import de.samply.template.ConverterTemplateUtils;
 
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Stream;
@@ -30,20 +25,20 @@ public abstract class TaggedLinesExplorer extends ExplorerImpl {
     }
 
     @Override
-    protected List<String> fetchLines(Path source, Pivot pivot) throws ExplorerException {
+    protected List<String> fetchLines(Path source, Pivot[] pivots) throws ExplorerException {
         try {
-            return fetchLinesWithoutExceptionHandling(source, pivot);
+            return fetchLinesWithoutExceptionHandling(source, pivots);
         } catch (IOException e) {
             throw new ExplorerException(e);
         }
     }
 
-    private List<String> fetchLinesWithoutExceptionHandling(Path source, Pivot pivot) throws IOException {
+    private List<String> fetchLinesWithoutExceptionHandling(Path source, Pivot[] pivots) throws IOException {
         try (Stream<String> lines = Files.lines(source, csvConfig.charset())) {
             List<String> resultLines = new ArrayList<>();
             AtomicInteger lineNumber = new AtomicInteger(1);
             AtomicBoolean isFirstElement = new AtomicBoolean(true);
-            lines.filter(line -> isLineToBeIncluded(line, lineNumber.getAndIncrement(), pivot)).forEach(line -> {
+            lines.filter(line -> isLineToBeIncluded(line, lineNumber.getAndIncrement(), pivots)).forEach(line -> {
                 line = editLine(line, lineNumber.get(), isFirstElement.get());
                 if (lineNumber.get() > 2 && isFirstElement.get()) {
                     isFirstElement.set(false);
@@ -54,38 +49,60 @@ public abstract class TaggedLinesExplorer extends ExplorerImpl {
         }
     }
 
-    private boolean isLineToBeIncluded(String line, int lineNumber, Pivot pivot) {
-        return (lineNumber == 1) || (hasLinePivot(line, pivot)) || isLastLine(line);
+    private boolean isLineToBeIncluded(String line, int lineNumber, Pivot[] pivots) {
+        return (lineNumber == 1) || (hasLinePivot(line, pivots)) || isLastLine(line);
     }
 
-    private boolean hasLinePivot(String line, Pivot pivot) {
-        if (!line.contains(pivot.attribute())) {
+    private boolean hasLinePivot(String line, Pivot[] pivots) {
+        Pivot tempPivot = Arrays.stream(pivots).filter(pivot -> {
+            if (line.contains(pivot.attribute())) {
+                Optional<String> value = fetchPivotValueFromLine(pivot.attribute(), line);
+                return !value.isEmpty() && value.get().equals(pivot.value());
+            }
             return false;
-        }
-        Optional<String> value = fetchPivotValueFromLine(pivot.attribute(), line);
-        return !value.isEmpty() && value.get().equals(pivot.value());
+        }).findFirst().orElse(null);
+        return Objects.nonNull(tempPivot);
     }
 
     @Override
-    public Optional<Pivot> fetchPivot(Path source, String pivotAttribute, int counter) throws ExplorerException {
+    public Optional<Pivot[]> fetchPivot(Path source, String pivotAttribute, int pageCounter, int pageSize) throws ExplorerException {
         try {
-            return fetchPivotWithoutExceptionHandling(source, pivotAttribute, counter);
+            return fetchPivotWithoutExceptionHandling(source, pivotAttribute, pageCounter, pageSize);
         } catch (IOException e) {
             throw new ExplorerException(e);
         }
     }
 
-    private Optional<Pivot> fetchPivotWithoutExceptionHandling(Path source, String pivotAttribute, int counter) throws IOException {
+    private Optional<Pivot[]> fetchPivotWithoutExceptionHandling(Path source, String pivotAttribute, int pageCounter, int pageSize) throws IOException {
         try (Stream<String> lines = Files.lines(source, csvConfig.charset())) {
             AtomicInteger tempCounter = new AtomicInteger(0);
-            String pivotLine = lines.filter(line -> tempCounter.getAndIncrement() == counter).findFirst().orElse(null);
-            if (pivotLine != null) {
+            List<Pivot> results = new ArrayList<>();
+            lines.filter(line -> {
+                int counter = tempCounter.getAndIncrement();
+                return counter >= pageCounter && counter < pageCounter + pageSize;
+            }).forEach(pivotLine -> {
                 Optional<String> pivotValue = fetchPivotValueFromLine(pivotAttribute, pivotLine);
                 if (pivotValue.isPresent()) {
-                    return Optional.of(new Pivot(pivotAttribute, pivotValue.get()));
+                    results.add(new Pivot(pivotAttribute, pivotValue.get()));
                 }
-            }
-            return Optional.empty();
+            });
+            return Optional.of(results.toArray(new Pivot[0]));
+        }
+    }
+
+    @Override
+    public int fetchTotalNumberOfElements(Path source) throws ExplorerException {
+        try {
+            return fetchTotalNumberOfElementsWithoutExceptionHandling(source);
+        } catch (IOException e) {
+            throw new ExplorerException(e);
+        }
+    }
+
+    private int fetchTotalNumberOfElementsWithoutExceptionHandling(Path source) throws IOException {
+        try (Stream<String> fileStream = Files.lines(source)) {
+            int numberOfLines = (int) fileStream.count();
+            return (numberOfLines >= 2) ? numberOfLines - 2 : 0;
         }
     }